// src/services/api/models/address.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Address
 */
export interface Address {
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    fullName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    phone?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    street1?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    street2?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    city?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    state?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    zip?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    country?: string | null;
}