// src/services/api/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LoginProvider } from './login-provider';
import { UserAttributes } from './user-attributes';
import { UserRole } from './user-role';
/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {string}
     * @memberof User
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof User
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof User
     */
    updated?: Date;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    password?: string | null;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    uid?: number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof User
     */
    termsAndPrivacyAgreedDate?: Date;
    /**
     * 
     * @type {LoginProvider}
     * @memberof User
     */
    loginProvider?: LoginProvider;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    isForgotPasswordRequested?: boolean;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    bio?: string | null;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    profilePicUrl?: string | null;
    /**
     * 
     * @type {UserRole}
     * @memberof User
     */
    role?: UserRole;
    /**
     * 
     * @type {UserAttributes}
     * @memberof User
     */
    userAttributes?: UserAttributes;
}